% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSpectra.R
\name{getSpectra}
\alias{getSpectra}
\title{Get spectra for a given samplecode}
\usage{
getSpectra(db, samplecode, spectratypes = c("reflectance",
  "pseudo-absorbance"))
}
\arguments{
\item{samplecode}{\code{samplecode} from \code{results} table}

\item{spectratypes}{(character) Spectra types to be included. If \code{NULL}, use all spectra types.}
}
\description{
Get spectra for a given samplecode
}
