% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initModel.R
\name{initModel}
\alias{initModel}
\title{Initialize results table for model}
\usage{
initModel(specdb, modelname)
}
\arguments{
\item{specdb}{\code{dplyr} \code{src} object pointing to spectra database}

\item{modelname}{Name of model with which to initialize table}
}
\description{
Initialize results table for model
}
