% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runInversion.R
\name{runInversion}
\alias{runInversion}
\title{Run PROSPECT inversion}
\usage{
runInversion(db, samplecode, modelname, custom_invert_options = NULL, ...)
}
\arguments{
\item{samplecode}{\code{samplecode} from \code{results} table}

\item{modelname}{Name of model with which to initialize table}

\item{custom_invert_options}{List of invert_options to modify}

\item{...}{Arguments passed on to \code{PEcAnRTM::invert.auto}
\describe{
  \item{observed}{Vector, matrix, or data frame (coerced to matrix) of 
observed values. For spectral data, wavelengths are rows and spectra are 
columns. Dimensions must align with the output of `model`.}
  \item{invert.options}{Parameters related to inversion.}
  \item{return.samples}{Include full samples list in output. Default = 
\code{TRUE}.}
  \item{save.samples}{Save samples to file as the inversion proceeds (useful 
for debugging). If \code{NULL}, do not save samples. Default = \code{NULL}.}
  \item{quiet}{Suppress progress bar and status messages. Default=FALSE}
  \item{parallel}{Logical. Whether or not to run multiple chains in parallel 
on multiple cores (default = \code{TRUE}).}
  \item{parallel.cores}{Number of cores to use for parallelization. If 
\code{NULL}
(default), allocate one fewer than detected number of cores.}
  \item{parallel.output}{Filename (or '' for stdout) for printing parallel 
outputs. Use with caution. Default = \code{'/dev/null'}.}
}}
}
\description{
Run PROSPECT inversion
}
